Change datatype of reduced sr features in recipe 1

+++ madmex/recipes/__init__.py
@@ -2,6 +2,7 @@ import datacube
 from datacube.storage.storage import write_dataset_to_netcdf
 from datacube.storage import masking
 import xarray as xr
+import numpy as np
 
 # TODO: Not sure whether the dask imports belog here or to another script downstream
 import dask
@@ -38,7 +39,7 @@ def madmex_001(x, y, time, dask_chunks, nc_filename):
     sr_clear = sr.where(clear)
     sr_clear2 = sr_clear.drop('pixel_qa')
     # Run temporal reductions and rename DataArrays
-    sr_mean = sr_clear2.mean('time', keep_attrs=True)
+    sr_mean = sr_clear2.mean('time', keep_attrs=True, dtype=np.int16)
     sr_mean.rename({'blue': 'blue_mean',
                     'green': 'green_mean',
                     'red': 'red_mean',
@@ -46,7 +47,7 @@ def madmex_001(x, y, time, dask_chunks, nc_filename):
                     'swir1': 'swir1_mean',
                     'swir2': 'swir2_mean',
                     'ndvi': 'ndvi_mean'}, inplace=True)
-    sr_min = sr_clear2.min('time', keep_attrs=True)
+    sr_min = sr_clear2.min('time', keep_attrs=True, dtype=np.int16)
     sr_min.rename({'blue': 'blue_min',
                     'green': 'green_min',
                     'red': 'red_min',
@@ -54,7 +55,7 @@ def madmex_001(x, y, time, dask_chunks, nc_filename):
                     'swir1': 'swir1_min',
                     'swir2': 'swir2_min',
                     'ndvi': 'ndvi_min'}, inplace=True)
-    sr_max = sr_clear2.max('time', keep_attrs=True)
+    sr_max = sr_clear2.max('time', keep_attrs=True, dtype=np.int16)
     sr_max.rename({'blue': 'blue_max',
                     'green': 'green_max',
                     'red': 'red_max',
@@ -62,7 +63,7 @@ def madmex_001(x, y, time, dask_chunks, nc_filename):
                     'swir1': 'swir1_max',
                     'swir2': 'swir2_max',
                     'ndvi': 'ndvi_max'}, inplace=True)
-    sr_std = sr_clear2.std('time', keep_attrs=True)
+    sr_std = sr_clear2.std('time', keep_attrs=True, dtype=np.int16)
     sr_std.rename({'blue': 'blue_std',
                     'green': 'green_std',
                     'red': 'red_std',
